refactor(analytics): Refactor classes into their own files PE-2139
